Allow app override (#107)
This extracts the creation of the web application in the server side.
This allows for people hacking on things to re-use that server for other
things (not a good idea, that's why it's not an arg to `start_server()`.

I think in the future, we can consider making it more generic, to allow
anyone to put /sendspin / `on_client_connect()` on their aiohttp app.
For now, we don't support it, but we allow hacking it in.